+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2022-06-18 (v0.1.3)
 - fix irregular volitional polite forms
+- add exception for negative aru forms
 
 ## 2022-05-23 (v0.1.2)
 - fix missing ri particle